refactor: (WIP) workflow to upsert invoice records
